requirements.txt
Appium-Python-Client
selenium
pandas
matplotlib